requirements.txt
asciimatics==1.14.0
libtmux==0.15.10
PyYAML==6.0
tinydb==4.7.0
